Refactor, add methods
